chore(mypy): drop the 33 outstanding errors (#45)
Categorised the existing mypy output and addressed each cluster:

Returning Any (9). CoolProp / numpy / pygfunction return `Any`, so any
function declaring a concrete return type ate a no-any-return:
  - hx_fan: wrap the two scalar UA/power computations in `float(...)`
  - heat_transfer: TDMA `phi.tolist()` → `[float(x) for x in phi]`
  - g_function: `R_b * eta / np.tanh(eta)` and `pipe.local_borehole_thermal_resistance()` cast to float
  - uv_treatment: annotate the nested `_row_to_dict` helper with `dict[str, float]`

Optional unsafe (5). `GroundSourceHeatPump.analyze_dynamic` could read /
write through a `None` `hp_result` if the off-mode fallback `_calc_state`
also returned None. Guard that fallback with an inert default row
(`hp_is_on=False, converged=False, Q_bhe [W]=0.0`) so the BHE
superposition step never sees None, and tighten `hp_is_on` to `bool(...)`.

assignment (2). In WSHPB / GSHPB `_determine_hp_state`, the on-branch
assignment `perf = self._calc_state(...) ... if perf is None: perf = ...`
left mypy unable to narrow. Switch to a clearer
`perf_opt = self._calc_state(...)
 perf = perf_opt if perf_opt is not None else self._calc_off_state(...)`
pattern that drops the union annotation naturally.

arg-type (3). ASHP passed `self.eta_cmp_isen` (typed `float | Callable
| None`) directly into `calc_ref_state` (which only accepts `float |
Callable`). Substitute `1.0` when it is None. GSHP attribute type
widened to `float | Callable` to match what `_eval_eff` already handled.
GSHP `_compute_bhe_superposition` got an explicit `bool(...)` for
`hp_is_on`.

no-redef (15). ASHPB / ASHP `_calc_state` defines names like `cs`,
`result`, the flow-state floats in both the `if not is_active:` early-
return branch and the on-branch below it. mypy reads that as
redefinition because both branches carry inline annotations. Two
combined fixes:
  - Set `[tool.mypy] allow_redefinition = true` in pyproject.
  - Drop the annotations on the second binding sites (mypy now inherits
    the type from the first branch, and the comments explain why).

Verified end-to-end on the touched code paths:

  $ uv run mypy src/physics_hp
  Success: no issues found in 29 source files

  $ uv run pytest -q
  8 passed

  $ uv run ruff check src/physics_hp
  All checks passed!

  $ uv run sphinx-build -W --keep-going -b html docs/source docs/build/html
  build succeeded.

Closes #37

+++ src/physics_hp/g_function.py
@@ -536,7 +536,7 @@ def calc_effective_borehole_thermal_resistance(
         eta = (H / (m_flow_pipe * cp_f)) * (1.0 / (2.0 * R_b)) * np.sqrt(1.0 + (4.0 * R_b) / R_a)
         if eta < 1e-6:
             return R_b
-        return R_b * eta / np.tanh(eta)
+        return float(R_b * eta / np.tanh(eta))
 
     elif boundary_condition == "uniform_heat_flux":
         # Hellström approximation for uniform heat flux
@@ -645,4 +645,4 @@ def calc_submerged_coil_thermal_resistance(
     pos = [(-D_s, 0.0), (D_s, 0.0)]
     pipe = gt.pipes.SingleUTube(pos, r_in, r_out, borehole, k_s_trick, k_g_trick, R_fp_total)
 
-    return pipe.local_borehole_thermal_resistance()
+    return float(pipe.local_borehole_thermal_resistance())
